requirements.txt
Flask==3.1.1
langchain==0.3.27
langchain_community==0.3.27
langchain_huggingface==0.3.1
langchain_pinecone==0.2.11
python-dotenv==1.1.1
setuptools==80.9.0
pypdf==5.6.0

-e .